fix type of lat lng

+++ database/migrations/2022_12_06_100659_stations_add_lat_lng.php
@@ -15,7 +15,7 @@ class StationsAddLatLng extends Migration
     {
         Schema::table('stations', function (Blueprint $table) {
             $table->decimal('lat', 8, 6);
-            $table->float('lng', 9, 6);
+            $table->decimal('lng', 9, 6);
         });
     }
